refactor: move functions
Move functions to make scope clearer and easier to find
