+++ go-chaos/internal/statefulset.go
@@ -18,15 +18,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"time"
+
 	v1 "k8s.io/api/apps/v1"
-	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"k8s.io/client-go/util/retry"
-	"strings"
-	"time"
 )
 
 func (c K8Client) GetZeebeStatefulSet() (*v1.StatefulSet, error) {
@@ -86,30 +83,4 @@ func (c K8Client) ScaleZeebeCluster(replicas int) (int, error) {
 	)
 
 	return initialReplicas, err
-}
-
-func (c K8Client) PauseReconciliation() error {
-	return c.setPauseFlag(true)
-}
-
-func (c K8Client) ResumeReconciliation() error {
-	return c.setPauseFlag(false)
-}
-
-func (c K8Client) setPauseFlag(pauseEnabled bool) error {
-	ctx := context.TODO()
-	namespace := c.GetCurrentNamespace()
-	clusterId := strings.TrimSuffix(namespace, "-zeebe")
-	zeebeCrd := schema.GroupVersionResource{Group: "cloud.camunda.io", Version: "v1alpha1", Resource: "zeebeclusters"}
-	payload := fmt.Sprintf(`{"metadata": {"labels": {"cloud.camunda.io/pauseReconciliation": "%t"}}}`, pauseEnabled)
-	err := retry.RetryOnConflict(retry.DefaultBackoff, func() error {
-		_, err := c.DynamicClient.Resource(zeebeCrd).Patch(ctx, clusterId, types.MergePatchType, []byte(payload), meta.PatchOptions{})
-		return err
-	})
-	if k8sErrors.IsNotFound(err) {
-		// No zb resource found so probably not Saas. Ignore for now.
-		fmt.Printf("Did not find zeebe cluster to pause reconciliation, ignoring. %s\n", err)
-		return nil
-	}
-	return err
 }
